refactor: moved lambda handler file location and removed unnecessary code in dbconn
